6327 | 8 files | M DumpSource2/schemas/client/CDOTA_Modifier_ArcWarden_Flux.h,  M DumpSource2/schemas/client/CDOTA_Modifier_Item_Fallen_Sky_Land.h,  M DumpSource2/schemas/server/CDOTA_Modifier_ArcWarden_Flux.h,  M DumpSource2/schemas/server/CDOTA_Modifier_Item_Fallen_Sky_Land.h,  M game/bin/built_from_cl.txt,  M game/dota/bin/linuxsteamrt64/libclient_strings.txt,  M game/dota/pak01_dir.txt,  M game/dota/steam.inf
https://steamdb.info/patchnotes/17433194/

+++ DumpSource2/schemas/client/CDOTA_Modifier_Item_Fallen_Sky_Land.h
@@ -1,6 +1,6 @@
 class CDOTA_Modifier_Item_Fallen_Sky_Land
 {
-	int32 impact_radius;
+	float32 impact_radius;
 	float32 land_time;
 	float32 burn_duration;
 };
